backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
pyjwt==2.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
uvicorn==0.24.0
pytest==7.4.3
httpx==0.25.2